==> notebooks/5_1_reproduce_fine_grained.ipynb <==
import sys,os
sys.path.append("../")

import pandas as pd
import numpy as np
from sklearn.cluster import KMeans
import matplotlib.pyplot as plt
from IPython.display import HTML
%matplotlib inline
from ipywidgets import interact, interactive, fixed, interact_manual
import ipywidgets as widgets
from scipy import stats
from sklearn.cluster import DBSCAN
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import MultiLabelBinarizer

r_thrshld = 5
k_thrshld = 5

## reading the change object and clustering.

article_name = "Truth"
change_object_dir =  "../data/change objects/"
filename =  f"{article_name}_change.h5"

change_object_file = os.path.join(change_object_dir, filename)

if os.path.exists(change_object_file):
    with pd.HDFStore(change_object_file, 'r') as store:
        change_object_dataframe = store.get("data")
else:
    print("file do not exist")
change_object_dataframe.shape

## Reduce the change object.

inner_join_change_object = change_object_dataframe[(change_object_dataframe["ins_tokens"]!=()) & (change_object_dataframe["del_tokens"]!=())]
display(inner_join_change_object.shape)
inner_join_change_object.head(2)

### Removing change object with insert or delete token size more than five.

change_object_dataframe["ins_length"]= change_object_dataframe["ins_tokens"].apply(lambda x: len(x))
change_object_dataframe["del_length"]= change_object_dataframe["del_tokens"].apply(lambda x: len(x))
change_object_dataframe.head(2)

bykau_change_object = inner_join_change_object[~((inner_join_change_object["ins_length"] >5 ) | (inner_join_change_object["del_length"] >5) )]
bykau_change_object.shape

bykau_change_object.reset_index(drop=True)[["ins_tokens", "del_tokens"]].head(3)

### Removing low user support tokens

# bykau_change_object = bykau_change_object.reset_index()
bykau_change_object = bykau_change_object.groupby("ins_tokens").filter(lambda x : x.index.get_level_values("editor").nunique()>=2)
bykau_change_object = bykau_change_object.groupby("del_tokens").filter(lambda x : x.index.get_level_values("editor").nunique()>=2)
bykau_change_object.shape

bykau_change_object["left_string"] = bykau_change_object["left_token"].str.join(" ")
bykau_change_object["ins_string"] = bykau_change_object["ins_tokens"].str.join(" ")
bykau_change_object["del_string"] = bykau_change_object["del_tokens"].str.join(" ")
bykau_change_object["right_string"] = bykau_change_object["right_token"].str.join(" ")

### Clustering using jaccard

##### Defining jaccard similarity function.

edit_tokens = bykau_change_object["ins_tokens"] + bykau_change_object["del_tokens"]
left_neighbours = bykau_change_object["left_token"].apply(lambda x: x[-r_thrshld:])
right_neighbours = bykau_change_object["right_token"].apply(lambda x: x[:r_thrshld])
neighbour_tokens = left_neighbours + right_neighbours
bykau_change_object["edit_tokens"] = edit_tokens.apply(lambda x: np.unique(x))
bykau_change_object["neighbour_tokens"] = neighbour_tokens.apply(lambda x: np.unique(x))
neighbour_vec=MultiLabelBinarizer().fit_transform(bykau_change_object["neighbour_tokens"])

db = DBSCAN(eps=1.5, min_samples=5, metric='jaccard').fit(neighbour_vec)
bykau_change_object["neighbour_clusters"] = db.labels_

repers_first = bykau_change_object.groupby("neighbour_clusters")[["left_string","del_string", "ins_string", "right_string"]].apply(lambda x: x.style.render())
# bykau_change_object.head(2)

@interact( clusters_html=fixed(repers_first), group=range(bykau_change_object.groupby("neighbour_clusters").ngroups))
def display_clusters(clusters_html, group):
     return display(HTML(clusters_html.iloc[group]))

# clustered_bykau_change_object = bykau_change_object[bykau_change_object["neighbour_clusters"]!= -1]
# group_edit_tokens = clustered_bykau_change_object.groupby("neighbour_clusters")["edit_tokens"].apply(lambda x: set(np.concatenate(x.values.ravel(), axis=0)))
# context_vec = MultiLabelBinarizer().fit_transform(group_edit_tokens)
# group_edit_dataframe = pd.DataFrame(group_edit_tokens.reset_index())

# db = DBSCAN(eps=0.8, min_samples=2, metric='jaccard').fit(context_vec)
# group_edit_dataframe["reclustered_group"] = db.labels_
# group_edit_dataframe = group_edit_dataframe[group_edit_dataframe["reclustered_group"]!= -1]



# final_dataframe = pd.merge( clustered_bykau_change_object, group_edit_dataframe[["reclustered_group", "neighbour_clusters"]], how="left",on="neighbour_clusters")

# final_dataframe.index = clustered_bykau_change_object.index

#### Saving change object and its clusters

bykau_dir =  "../data/bykau_change_object/"
filename =  f"{article_name}_change.h5"

change_object_file = os.path.join(bykau_dir, filename)
with pd.HDFStore(change_object_file, 'w') as store:
    store.put("data", final_dataframe,)

### Grouping and showing clusters

final_dataframe["left_string"] = final_dataframe["left_token"].str.join(" ")
final_dataframe["ins_string"] = final_dataframe["ins_tokens"].str.join(" ")
final_dataframe["del_string"] = final_dataframe["del_tokens"].str.join(" ")
final_dataframe["right_string"] = final_dataframe["right_token"].str.join(" ")

final_dataframe_grouped = final_dataframe.groupby("reclustered_group")

repers = final_dataframe_grouped[["left_string","del_string", "ins_string", "right_string"]].apply(lambda x: x.style.render())

# all_html = " ".join(repers_4_full)
# file_name = article_name + "_4_full_"+str(NO_OF_CLUSTERS) + "_clusters.html"
# file_path = os.path.join("./visualisation", file_name)
# with open(file_path, 'wb') as f:
#     f.write(all_html.encode())

## Vaisualisation of clusters
##### Rerun next cell each time this notebook is reloaded, this is a bug in current jupyterlab.

@interact( clusters_html=fixed(repers), group=range(final_dataframe_grouped.ngroups))
def display_clusters(clusters_html, group):
     return display(HTML(clusters_html.iloc[group]))
